feat(2023/11): solve in Ruby and by dimension
The basic idea is, since the "expandability" and the manhattan distance can be both split up into the two dimensions X/Y independently, to only collect the list of X and Y coordinates of each "galaxy", and process those lists independently and in linear time. As the only non-linear thing we have to sort the X list once. We can then make use of the recursive structure to walk the both lists only once and sum up the required numbers of fixed and expandable spots. A diagram like this helps:

```
 1
 -****6
      -**
 -****-**9
         -****
      -**-****
 -****-**-****14
              -*****
         -****-*****
      -**-****-*****
 -****-**-****-*****20
              -*****
         -****-*****
      -**-****-*****
 -****-**-****-*****20
```

We basically have to count the number of fixed ("-") and expandable ("*") spots here, and notice the number of spots added for the next entry is the number of entries added for the previous entry, plus some number dependent on the distance to the previous entry and the number of entries seen so far.

We can do this for X and Y independently and add the values accordingly.